Update index.html
With this update, the updateRecommendations() function filters products whose integer values, when added together, closely align with the result value obtained from the result element. The product list will now display products that can be combined to achieve the calculated result value as closely as possible. If no products match the criteria, a message indicating so will be displayed.